renamed LoginBanner-> LoginForm
